processor/requirements.txt
pika
pillow
scikit-learn
numpy
requests